feat: split off test printer socket store
